Delete task not working

+++ Controllers/ProjectsController.cs
@@ -155,6 +155,8 @@ namespace TaskTracker.Controllers
         // to grab the id from the URL. It is then made available to us as the `id` argument to the method.
         //
         [HttpDelete("{id}")]
+        [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
+
         public async Task<IActionResult> DeleteProject(int id)
         {
             // Find this project by looking for the specific id
@@ -164,6 +166,20 @@ namespace TaskTracker.Controllers
                 // There wasn't a project with that id so return a `404` not found
                 return NotFound();
             }
+
+            if (project.UserId != GetCurrentUserId())
+            {
+                // Make a custom error response
+                var response = new
+                {
+                    status = 401,
+                    errors = new List<string>() { "Not Authorized" }
+                };
+
+                // Return our error with the custom response
+                return Unauthorized(response);
+            }
+
 
             // Tell the database we want to remove this record
             _context.Projects.Remove(project);
